requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
python-dotenv==1.0.1
tqdm

numpy==2.1.1
pandas==2.2.3
geopandas==1.0.1
scikit-learn==1.5.2
xgboost==2.1.1
torch==2.4.1
ollama
sentence_transformers
transformers

geopandas==1.0.1
shapely==2.0.6
pyarrow==17.0.0

jupyter==1.1.1
notebook==7.2.2
matplotlib==3.9.2

pytest==8.3.3
httpx==0.27.2
pyyaml